<UI> Weight, Condition Information
Weight and Conditions now displayed.
Redesigned textassociation and myfloat to give the rounding responsibility to textassociation rather than myfloat. This is so that a) operations can be performed on the raw values and b) where multiple text associations are used for the same variable, different precisions can be displayed in each text element.