<Feature> "Working" Inventory System
As far as I have tested the inventory system works.
Inventory items can be stacked and moved between inventories.
Duplicate item objects do not exist in the same inventory unless the
item is not stackable (for example to preserve weapon data, which would
otherwise be lost).
Stackable items should be used when items are identical e.g. resources.
Items can remain in the inventory if marked as don't destroy. This is
useful if you want to keep track of resources in a hub area or (for
example) character money, which can be 0 but is still tracked in the
inventory. Other items are destroyed if their quantity reaches 0.