<Refactoring> NONFUNCTIONAL - SEE DETAILS
Lots of modifications:
Home is now an inventory which contains Characters, as well as Resources
and Items.
Inventories are responsible for saving themselves, this should make it
easy to cascade Save and Load calls down through inventory items and
return the world state to how it used to be.
Nailing down the functionality of the MenuList, increasing functionality
for PopUp windows.
Also figuring out what UI classes are needed for the MenuLists, and
which Menus actually use these lists.
Reassessed what should and shouldn't be a MyGameObject- Regions,
Inventories, and States are now resources. This means a StateMachine
could/should be a kind of inventory. The cascading save sytem should
also make it easier to Save and Load statemachine data.