<Feature> Canyon
Split the various region generation algorithms into Monobehaviours for
each region type.
Added a canyon generator that creates a long thin region with a small
number of rocks.
Added a durability/life bar for weapons.
Did some quick sprite ideas for plants.
Improved area generation so it only has to perform the collision check
once when entering a region for the first time.
Wrote some lore for the universe.