<Feature> Ruins
Ruin generation works reasonably nicely now.
Some shapes are generated the wrong way round, so there is a bug in the
triangulator code.
Ruin walls are combined into a single mesh.
Still needs corners to be calculated when there is a 270* angle between
walls.
Also needs some slight perturbation to be applied to the walls to make
them look less smooth.